Move webservices implementation into pure gosu (yikes, we need tools)
First pass at a pure gosu implementation exposing webservices.  You can get a list of available webservices at

  localhost:8080/webservices

and add new webservices by creating a class in the 'webservices' package with an @gw.xml.ws.annotation.WsiWebService annotation on it.

Performance is atrocious, and I discovered a bug when you implicitly implement a method in an interface with a superclass in java.

Yay.